refactor(jobsprovider): function to create new job
